First round of refactoring item versions: better permissions and vesrioning module for utilities
