+++ pydatalab/src/pydatalab/models/__init__.py
@@ -7,13 +7,7 @@ from pydatalab.models.files import File
 from pydatalab.models.people import Person
 from pydatalab.models.samples import Sample
 from pydatalab.models.starting_materials import StartingMaterial
-from pydatalab.models.versions import (
-    CompareVersionsQuery,
-    ItemVersion,
-    RestoreVersionRequest,
-    VersionAction,
-    VersionCounter,
-)
+from pydatalab.models.versions import ItemVersion
 
 ITEM_MODELS: dict[str, type[BaseModel]] = {
     "samples": Sample,
@@ -30,11 +24,6 @@ __all__ = (
     "Cell",
     "Collection",
     "Equipment",
-    "ITEM_MODELS",
     "ItemVersion",
-    "VersionCounter",
-    "UserSnapshot",
-    "VersionAction",
-    "RestoreVersionRequest",
-    "CompareVersionsQuery",
+    "ITEM_MODELS",
 )
